requirements.txt
python-telegram-bot==13.7
openai==0.28
google-cloud-speech